requirements.txt
aadict==0.2.3
asgiref==3.5.2
asset==0.6.13
async-timeout==4.0.2
autopep8==1.7.0
cachelib==0.9.0
click==8.1.2
colorama==0.4.5
cssbeautifier==1.14.6
cssmin==0.2.0
Deprecated==1.2.13
djlint==1.12.3
dnspython==2.2.1
EditorConfig==0.12.3
email-validator==1.1.3
Flask==2.1.1
Flask-Assets==2.0
Flask-Login==0.6.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
globre==0.1.5
greenlet==1.1.2
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jsbeautifier==1.14.6
jsmin==3.0.1
lessc==0.1.3
MarkupSafe==2.1.1
netaddr==0.8.0
packaging==21.3
pathspec==0.10.1
pkg_resources==0.0.0
pycodestyle==2.9.1
PyMySQL==1.0.2
pyparsing==3.0.9
pypinyin==0.47.1
python-dotenv==0.20.0
PyYAML==6.0
redis==4.3.4
regex==2022.8.17
six==1.16.0
SQLAlchemy==1.4.35
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
uWSGI==2.0.20
webassets==2.0
Werkzeug==2.1.1
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
